move tests to their own folder
